# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DevSignInRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "DevSignInRequest",
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable error codes. Values are part of the public API contract.",
        "enum": [
          "validation_failed",
          "unauthenticated",
          "forbidden",
          "not_found",
          "conflict",
          "idempotency_key_reused",
          "payload_too_large",
          "dependency_unavailable",
          "internal_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "description": "Everything the console needs to render its shell after sign-in.",
        "properties": {
          "organisations": {
            "items": {
              "$ref": "#/components/schemas/OrganisationMembershipResponse"
            },
            "title": "Organisations",
            "type": "array"
          },
          "principal_kind": {
            "title": "Principal Kind",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the caller is a service account."
          }
        },
        "required": [
          "principal_kind",
          "organisations"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "OrganisationMembershipResponse": {
        "properties": {
          "organisation": {
            "$ref": "#/components/schemas/OrganisationResponse"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "The caller's role in this organisation."
          }
        },
        "required": [
          "organisation",
          "role"
        ],
        "title": "OrganisationMembershipResponse",
        "type": "object"
      },
      "OrganisationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "created_at"
        ],
        "title": "OrganisationResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "The body returned for every non-2xx API response.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "correlation_id": {
            "description": "Quote this when reporting a failure.",
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "description": "Structured, non-sensitive context such as failing field paths.",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human-readable summary. Safe to display.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "Role": {
        "description": "A member's role within one organisation.",
        "enum": [
          "owner",
          "admin",
          "developer",
          "reviewer",
          "viewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AgentRail platform API.\n\nPhase 0 exposes the deterministic job slice used to prove the end-to-end request\npath: the API records a job in PostgreSQL, publishes its identifier to Redis, and\na worker executes it against the CloudOps sandbox.",
    "title": "AgentRail API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/dev/session": {
      "post": {
        "description": "Deterministic sign-in used by local development, CI and the demo.\n\nUnavailable in staging and production: :meth:`ApiSettings.dev_auth_enabled`\nis false there, and this route reports 404 rather than 403 so its existence\nis not advertised.",
        "operationId": "dev_sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Disabled in this environment."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign in without a provider (non-deployed environments only)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/github/callback": {
      "get": {
        "operationId": "github_callback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete GitHub sign-in",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "read_me",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The current caller and their organisations",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}